scripts/requirements.txt
pandas==2.0.3
numpy==1.24.4
psycopg2-binary==2.9.6
matplotlib==3.7.2 